+++ R/convenience-plots.R
@@ -4,7 +4,7 @@
 #' smooth terms contained in the model. If more than one smooth is present, the
 #' different smooth are faceted.
 #'
-#' @param x A data frame or object of class \code{pamm}.
+#' @param x A data frame or object of class \code{ped}.
 #' @param ... Further arguments passed to \code{\link{get_terms}}
 #' @import ggplot2
 #' @return A \code{\link[ggplot2]{ggplot}} object.