+++ R/formula-specials.R
@@ -268,11 +268,7 @@ add_concurrent <- function(ped, data, id_var) {
     ccr_vars_i <- c(tz_var_i, tdc_vars_i)
     ccr_i_df   <- data %>%
       select(one_of(c(id_var, ccr_vars_i)))
-    if(tidyr_new_interface()) {
-      ccr_i_df <- ccr_i_df %>% unnest(cols = -one_of(id_var))
-    } else {
-      ccr_i_df <- ccr_i_df %>% unnest()
-    }
+    ccr_i_df <- ccr_i_df %>% unnest(cols = -one_of(id_var))
 
     li <- map2(ped_split, split(ccr_i_df, f = ccr_i_df[[id_var]]),
       function(.x, .y) {